pyproject.toml
[tool.poetry]
name = "python-open-restaurant-django"
version = "0.1.0"
description = ""
authors = ["Matthew Lonis <mattlonis@gmail.com>"]
readme = "README.md"
packages = [{include = "python_open_restaurant_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"

[tool.poetry.group.test.dependencies]
flake8 = "^6.0.0"
pylint = "^2.16.3"
pylint-django = "^2.5.3"
pylint-junit = "^0.3.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
configupdater = "^3.1.1"
isort = "^5.12.0"
pre-commit = "^3.1.1"
ruamel-yaml = "^0.17.21"
tomlkit = "^0.11.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 125

[tool.isort]
line_length = 125
profile = "black"

[tool.pytest.ini_options]
addopts = "--ignore=./utility-repo-scripts"
log_cli = false
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
